+++ src/emptyState/style.ts
@@ -1,4 +1,5 @@
 import css from 'styled-jsx/css'
+import { space } from '_utils/branding'
 
 export default css`
   .kirk-empty-state {
@@ -6,6 +7,11 @@ export default css`
     flex-direction: column;
     align-items: center;
     width: 100%;
+    text-align: center;
+  }
+
+  :global(.kirk-empty-state .kirk-title) {
+    margin: ${space.xl} 0;
   }
 
   .kirk-empty-state img {